4_movie_recommendation/notebooks/requirements.txt
# Core data science
pandas>=1.3.0
numpy>=1.21.0

# Machine learning
scikit-learn>=1.0.0
scikit-surprise>=1.1.1

# Web interface
streamlit>=1.12.0

# Optional: for saving/loading models
pickle5; python_version < "3.8"